backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
pgvector>=0.3.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
openai>=1.50.0
google-generativeai>=0.8.0
faster-whisper>=1.0.3
ffmpeg-python>=0.2.0
sentence-transformers>=3.1.1
torch>=2.5.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0
google-genai>=1.65.0
